+++ players/random_agent.py
@@ -1,0 +1,12 @@
+from agent import Agent
+
+class RandomAgent(Agent):
+    """
+    
+    """
+    def __init__(self, name="Random Agent"):
+        super().__init__(name)
+    
+    def choose_action(self, state):
+        return action
+
